+++ packages/dara-components/tests/python/common/test_base_component.py
@@ -32,9 +32,6 @@ class TestBaseComponent(unittest.TestCase):
                 'children': [t1.dict(exclude_none=True), t2.dict(exclude_none=True), t1.dict(exclude_none=True)],
                 'height': '10%',
                 'width': '10px',
-                'bold': False,
-                'italic': False,
-                'underline': False,
             },
             'uid': 'uid',
         }