+++ packages/dara-components/tests/python/common/test_form_page.py
@@ -20,9 +20,6 @@ class TestFormPageComponent(unittest.TestCase):
             'name': 'FormPage',
             'props': {
                 'children': [s1.dict(exclude_none=True)],
-                'bold': False,
-                'italic': False,
-                'underline': False,
                 'title': 'Page',
             },
             'uid': str(test_uid),