+++ packages/dara-components/tests/python/common/test_icon.py
@@ -15,9 +15,6 @@ class TestIconComponent(unittest.TestCase):
             'name': 'Icon',
             'props': {
                 'children': [],
-                'bold': False,
-                'italic': False,
-                'underline': False,
                 'icon': 'wrench',
                 'color': 'red',
             },