+++ packages/dara-components/tests/python/smart/test_code_editor.py
@@ -20,9 +20,6 @@ class TestCodeEditorComponent(unittest.TestCase):
             'name': 'CodeEditor',
             'props': {
                 'script': script.dict(exclude_none=True),
-                'bold': False,
-                'italic': False,
-                'underline': False,
             },
             'uid': str(test_uid),
         }